+++ src/core/models/container.js
@@ -408,7 +408,7 @@ class Container extends Syncable {
 
   trigger(evtName, args) {
     this._clearObject();
-    super.trigger(evtName, args);
+    return super.trigger(evtName, args);
   }
 
   /**