+++ src/ui/messages/layer-dialog-message-view-container.js
@@ -80,7 +80,7 @@ registerComponent('layer-dialog-message-view-container', {
     },
 
     onAfterCreate() {
-      this.model.on('change', this.onRerender, this);
+      this.model.on('message-type-model:change', this.onRerender, this);
     },
 
     /**